chore: rename alert channel template file
