+++ src/templates/alert-channel.js
